refactor add player to game to be method of the game class instead of the game manager

+++ src/server/internal/player/player.go
@@ -3,7 +3,7 @@ package player
 import "github.com/gorilla/websocket"
 
 type Player struct {
-	Id         uint8
+	Id         uint64
 	SnakeColor SnakeColor
 	Score      int
 	GameId     string
